refactor loop handling into own machine
Add another layer of state machines with to represent the loop cycling itself.
Making this it's own component allows the possibility for the driver to use
the builder to register multiple loops.  This patch is intended to provide the
decoupling without changing any behavior, there is is only one loop.  With
this patch, there is clear insight that at least both SESSION_RESOLVED and
SESSION_ESTABLISHED need to be potentially "broadcast" to multiple recipients.