capture should collect details about each frame
Without this patch, xstate never executes startCapture, because it's listed in
the after section without a delay, which is an invalid spec.  As a result,
there is a bug where each capture file includes the cumulative history of all
previous traces.
This patch fixes the bug so that the capture is reset after recording it.  This
allows the capture for each cycle to include the logs for requests during that
cycle and only that cycle.