discern between GUARDIAN M2M vs BLE endpoint
The "monitor data" endpoint contains information about the deviceFamily.  The
deviceFamily determines which endpoint will be responsible for data about the
device. When the deviceFamily is GUARDIAN, the M2M endpoint is used.
In fact, when the deviceFamily is GUARDIAN, the blePereodicEndpoint responds
with an empty string.

In the US, we've other devices where, deviceFamily is BLE_X.  For these devices,
M2M still returns data, but with empty stubs so that the effect is a no-op
pass-through.  For examples, it has a data structure for glucose values, with
no real values. In this case, the medicalDeviceFamily does not match the
deviceFamily, and is set to PARADIGM, with kind set to Main, and deviceKind
set to null.
However, the blePereodicEndpoint responds with a data structure with medicalDeviceFamily set to NGP, kind set to BLENGP, with the actual device history data attached.

* https://github.com/NightscoutFoundation/xDrip/blob/3fc4c1634415ce96d874b67ee6e4049fc45be732/app/src/main/java/com/eveningoutpost/dexdrip/cgm/carelinkfollow/client/CareLinkClient.java#L252

With this patch, both style of devices work.  In the event where devices that
are GUARDIAN devices need to use the BLE endpoint, or in the event where
devices that are not GUARDIAN devices need to use need to use the M2M endpoint
additional work will be required.