robust internal handling for handling last known specs
This prevents a category of errors the plugin is running internally to
Nightscout but has no recent information to generate a reasonable last known
basis.
This allows minimed to sync treatments data specifically on last treatment data
when syncing from within Nightscout.

There is a potential issue with using 0 as the default here, as this could
create issues duplicating and syncing the largest amount of data that the
driver may allow.