minimed: promised error control flow
With a re-used session, it's possible that session authentication works but
nonetheless sometime later, the tokens are rejected when fetching data. In
these circumstances, rejecting the promise by returning Promise.reject( ) is
required in a pipelined or composed workflow.

In practice, this patch makes it possible to run forever more reliably by
recovering from anticipated failures.  Without this patch, refreshing the token
would appear to work, but eventually no data would be returned, or a promise
will fail resolve/reject correctly.  Rejecting the failures properly allows
further consumers of the data down the chain to consume the data with fewer
checks.